cosmetic_app/pyproject.toml
[tool.poetry]
name = "cosmetic-app"
version = "0.1.0"
description = ""
authors = ["tesseract <tesseractmaks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["full"], version = "^0.104.1"}
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
hypercorn = {extras = ["full"], version = "^0.15.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
pillow = "^10.1.0"
fastapi-pagination = "^0.12.13"
fastapi-filter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
